Fixed GuestController
